+++ packages/expo-cli/src/commands/utils/__tests__/extractTemplateAppAsync-test.ts
