+++ packages/expo-cli/src/commands/utils/extractTemplateAppAsync.ts
